agents/requirements.txt
hypercorn
restate_sdk==0.5.1
pydantic
openai-agents==v0.0.4